realtime-lipsync/tray/requirements.txt
pystray>=0.19
pillow>=9.0
opencv-python>=4.8
websockets>=12.0
numpy>=1.24
pyvirtualcam>=0.10
